requirements.txt
google-genai>=0.8.0
mcp>=1.0.0
httpx
pydantic>=2.0.0
python-dotenv
aiofiles
PyGithub
gitpython
